www_immortal/requirements.txt
django >= 2.1.6, < 3.0 
django-appconf==1.0.2
django-countries==5.3.1
django-image-cropping==1.2.0
django-ipware==2.1.0
django-select2==6.3.1
django-stdimage==3.2.0
django-tinymce==2.7.0
django-uuslug==1.1.8
easy-thumbnails==2.5
Pillow==5.2.0
pip==18.0
progressbar2==3.38.0
psycopg2==2.7.5
python-slugify==1.2.5
python-utils==2.3.0
pytz==2018.5
setuptools==39.0.1
six==1.11.0
Unidecode==1.0.22
